Move controller schema validation to a controller concern
